Move count_flot to coffee_mode.py
